extract out Storage class
