+++ src/main/java/duke/Duke.java
@@ -3,11 +3,7 @@ package duke;
 import duke.exceptions.EmptyCommandException;
 import duke.exceptions.UnknownCommandException;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Scanner;
 
 public class Duke {
@@ -26,30 +22,29 @@ public class Duke {
     public static final String EMPTY_COMMAND_MESSAGE = "Command is empty!";
     public static final String OUTPUT_FILE = "outputfile.txt";
     public static final String IOEXCEPTION_ERROR_MESSAGE = "IOException Error";
-    public static final String FILE_NOT_FOUND_MESSAGE = "File Not Found";
 
     private Ui ui;
     private Parser parser;
     private TaskList taskList;
+    private Storage storage;
 
     public Duke() {
         ui = new Ui();
         parser = new Parser();
+        storage = new Storage();
         taskList = new TaskList();
     }
 
     public void run() {
         ui.showWelcomeMessage();
-        load(OUTPUT_FILE);
+        taskList = storage.load(OUTPUT_FILE);
         starting();
         ui.showEndingMessage();
         System.exit(0);
     }
 
     public static void main(String[] args) {
-
         new Duke().run();
-
     }
 
     private void starting() {
@@ -77,7 +72,6 @@ public class Duke {
 
 
     public void processInput(String line) throws UnknownCommandException, IOException, EmptyCommandException {
-
         String[] words = line.split(" ", 2);
         String command = words[0];
         // words[0] is the command, words[n] is the next few words
@@ -86,42 +80,40 @@ public class Duke {
         case TODO_COMMAND:
             Todo td = taskList.createTodo(words);
             ui.printAddTaskMessage(td);
-            save(OUTPUT_FILE);
+            storage.save(taskList, OUTPUT_FILE);
             break;
         case DEADLINE_COMMAND:
             Deadline d = taskList.createDeadline(words);
             ui.printAddTaskMessage(d);
-            save(OUTPUT_FILE);
+            storage.save(taskList, OUTPUT_FILE);
             break;
         case EVENT_COMMAND:
             Event e = taskList.createEvent(words);
             ui.printAddTaskMessage(e);
-            save(OUTPUT_FILE);
+            storage.save(taskList, OUTPUT_FILE);
             break;
         case DELETE_COMMAND:
             deleteTask(words);
-            save(OUTPUT_FILE);
+            storage.save(taskList, OUTPUT_FILE);
             break;
         case MARK_COMMAND:
             int markIndex = Integer.parseInt(words[1]) - 1; // 0 indexing
             taskList.markIndexAsDone(markIndex);
             ui.printMarkedMessage(taskList.getTask(markIndex));
-            save(OUTPUT_FILE);
+            storage.save(taskList, OUTPUT_FILE);
             break;
         case UNMARK_COMMAND:
             int unmarkIndex = Integer.parseInt(words[1]) - 1; // 0 indexing
             taskList.unmarkIndexAsDone(unmarkIndex);
             ui.printUnmarkedMessage(taskList.getTask(unmarkIndex));
-            save(OUTPUT_FILE);
+            storage.save(taskList, OUTPUT_FILE);
             break;
         case LIST_COMMAND:
             taskList.printList();
             break;
-
         default:
             throw new UnknownCommandException();
         }
-
     }
 
     private void deleteTask(String[] words) {
@@ -139,92 +131,6 @@ public class Duke {
         if (words.length < 2 || words[1].equals("")) {
             throw new EmptyCommandException();
         }
-    }
-
-
-    private void writeTasksToFile(String filePath) throws IOException {
-        FileWriter fw = new FileWriter(filePath);
-        ArrayList<Task> tasks = taskList.getTaskList();
-        for (Task t : tasks) {
-            if (t != null) {
-                fw.write(t.saveFormat() + System.lineSeparator());
-            }
-        }
-
-        fw.close();
-    }
-
-    private void save(String filename) throws IOException {
-        try {
-            writeTasksToFile(filename);
-        } catch (IOException e) {
-            System.out.println("Something went wrong: " + e.getMessage());
-        }
-    }
-
-    private void load(String filename) {
-        File f = new File(filename);
-
-        if (f.exists()) {
-            try {
-                loadFileContents(f);
-            } catch (FileNotFoundException e) {
-                System.out.println(FILE_NOT_FOUND_MESSAGE);
-            }
-        }
-    }
-
-    private void loadFileContents(File f) throws FileNotFoundException {
-        Scanner s = new Scanner(f); // create a Scanner using the File as the source
-        while (s.hasNext()) {
-            processFileContents(s.nextLine());
-        }
-    }
-
-    private void processFileContents(String line) {
-        String[] words = line.split("\\|");
-        String type = words[0];
-        boolean isDone = false;
-        if (Integer.valueOf(words[1]) == 1) {
-            isDone = true;
-        }
-        switch (type) {
-        case "T":
-            loadTodo(words, isDone);
-            break;
-        case "E":
-            loadEvent(words, isDone);
-            break;
-        case "D":
-            loadDeadline(words, isDone);
-            break;
-        default:
-            // unknown char error
-        }
-    }
-
-    private void loadDeadline(String[] words, boolean isDone) {
-        Deadline deadline = new Deadline(words[2], words[3]);
-        if (isDone) {
-            deadline.markAsDone();
-        }
-        taskList.addDeadline(deadline);
-    }
-
-    private void loadEvent(String[] words, boolean isDone) {
-        Event event = new Event(words[2], words[3], words[4]);
-        if (isDone) {
-            event.markAsDone();
-        }
-        taskList.addEvent(event);
-    }
-
-    private void loadTodo(String[] words, boolean isDone) {
-        Todo todo = new Todo(words[2]);
-        if (isDone) {
-            todo.markAsDone();
-        }
-        taskList.addTodo(todo);
     }
